Refactoring the project and configuring Cap to can handle integration event

+++ src/Modules/Identity/Identity.Application/Commands/DeleteUser/DeleteUserHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Identity.Application.Abstractions.IdentityProvider;
+using Identity.Application.Errors;
 using Identity.Domain.Repositories;
 using MediatR;
 using Shared.Application.Abstractions.CQRS;
@@ -13,13 +14,25 @@ public class DeleteUserHandler(
 {
     public async Task<Result<Unit>> Handle(DeleteUserCommand command, CancellationToken cancellationToken)
     {
-        var user = await userRepository.GetById(command.UserId, cancellationToken)
-            ?? throw new ArgumentException("User Not Found");
+        try
+        {
+            var user = await userRepository.GetById(command.UserId, cancellationToken);
+        
+            if(user is null)
+                return Result<Unit>.Failure(UserErrors.UserNotFound);
 
-        await identityProvider.DeleteUser(user.KeycloakId, cancellationToken);
-        await userRepository.Delete(user, cancellationToken);
-        await userRepository.SaveChanges(cancellationToken);
+            await identityProvider.DeleteUser(user.KeycloakId, cancellationToken);
+            await userRepository.Delete(user, cancellationToken);
+            await userRepository.SaveChanges(cancellationToken);
 
-        return Result<Unit>.Success(Unit.Value);
+            return Result<Unit>.Success(Unit.Value);
+        }
+        catch (Exception ex)
+        {
+            return Result<Unit>.Failure(new Error(
+                "UserDelete.Error",
+                "DeleteUserError", 
+                ex.Message));
+        }
     }
 }
